Mock datetime and move prints a level up
