+++ runtime/runtime-params-estimator/estimator-warehouse/Cargo.toml
@@ -11,9 +11,9 @@ edition = "2021"
 anyhow = "1.0.56"
 chrono = "0.4.19"
 clap = {version = "3.1.6", features = ["derive"]}
-nix = "0.15.0" 
+nix = "0.15.0"
 reqwest = {version = "0.11.0", features = ["blocking"]}
-rusqlite = {version = "0.27.0", features = ["bundled", "chrono"]}
+rusqlite = {version = "0.27.0", features = ["bundled", "chrono", "functions"]}
 serde = {version = "1.0.136", features = ["derive"]}
 serde_json = "1.0.68"
 tempfile = "3.3.0"
@@ -21,4 +21,3 @@ xshell = "0.2.1"
 
 [dev-dependencies]
 insta = "1.13.0"
-regex = "1.5.5"
